+++ lib/rest/serverless/v1/service/build.d.ts
@@ -12,7 +12,7 @@ import { BuildStatusList } from './build/buildStatus';
 import { BuildStatusListInstance } from './build/buildStatus';
 import { SerializableClass } from '../../../../interfaces';
 
-type BuildRuntime = 'node8'|'node10'|'node12';
+type BuildRuntime = 'node8'|'node10'|'node12'|'node14';
 
 type BuildStatus = 'building'|'completed'|'failed';