+++ lib/rest/verify/v2/service/verification.d.ts
@@ -69,6 +69,8 @@ interface VerificationListInstance {
  * @property payee - The payee of the associated PSD2 compliant transaction
  * @property rateLimits - The custom key-value pairs of Programmable Rate Limits.
  * @property sendDigits - The digits to send after a phone call is answered
+ * @property templateCustomSubstitutions - The values of the special variables declared on the message template.
+ * @property templateSid - The verification template SMS messages.
  * @property to - The phone number or email to verify
  */
 interface VerificationListInstanceCreateOptions {
@@ -83,6 +85,8 @@ interface VerificationListInstanceCreateOptions {
   payee?: string;
   rateLimits?: object;
   sendDigits?: string;
+  templateCustomSubstitutions?: string;
+  templateSid?: string;
   to: string;
 }