+++ spec/integration/rest/messaging/v1/service/usAppToPersonUsecase.spec.js
@@ -80,8 +80,8 @@ describe('UsAppToPersonUsecase', function() {
                   'post_approval_required': false
               },
               {
-                  'code': 'CONVERSATIONAL',
-                  'name': 'Conversational Messaging',
+                  'code': 'PROXY',
+                  'name': 'Proxy',
                   'description': 'Peer-to-peer app-based group messaging with proxy/pooled numbers (For example: GroupMe)\nSupporting personalized services and non-exposure of personal numbers for enterprise or A2P communications. (Examples include: Uber and AirBnb.)',
                   'post_approval_required': true
               },
@@ -114,6 +114,12 @@ describe('UsAppToPersonUsecase', function() {
                   'name': 'Higher Education',
                   'description': 'For campaigns created on behalf of Colleges or Universities and will also include School Districts etc that fall outside of any \'free to the consumer\' messaging model',
                   'post_approval_required': false
+              },
+              {
+                  'code': 'K12_EDUCATION',
+                  'name': 'K-12 Education',
+                  'description': 'Campaigns created for messaging platforms that support schools from grades K-12 and distance learning centers. This is not for Post-Secondary schools.',
+                  'post_approval_required': true
               },
               {
                   'code': 'LOW_VOLUME',
@@ -213,8 +219,8 @@ describe('UsAppToPersonUsecase', function() {
                   'post_approval_required': false
               },
               {
-                  'code': 'CONVERSATIONAL',
-                  'name': 'Conversational Messaging',
+                  'code': 'PROXY',
+                  'name': 'Proxy',
                   'description': 'Peer-to-peer app-based group messaging with proxy/pooled numbers (For example: GroupMe)\nSupporting personalized services and non-exposure of personal numbers for enterprise or A2P communications. (Examples include: Uber and AirBnb.)',
                   'post_approval_required': true
               },
@@ -247,6 +253,12 @@ describe('UsAppToPersonUsecase', function() {
                   'name': 'Higher Education',
                   'description': 'For campaigns created on behalf of Colleges or Universities and will also include School Districts etc that fall outside of any \'free to the consumer\' messaging model',
                   'post_approval_required': false
+              },
+              {
+                  'code': 'K12_EDUCATION',
+                  'name': 'K-12 Education',
+                  'description': 'Campaigns created for messaging platforms that support schools from grades K-12 and distance learning centers. This is not for Post-Secondary schools.',
+                  'post_approval_required': true
               },
               {
                   'code': 'LOW_VOLUME',