+++ spec/integration/rest/serverless/v1/service/build.spec.js
@@ -139,7 +139,7 @@ describe('Build', function() {
                   'version': '1.0.1'
               }
           ],
-          'runtime': 'node10',
+          'runtime': 'node14',
           'status': 'building',
           'date_created': '2018-11-10T20:00:00Z',
           'date_updated': '2018-11-10T20:00:00Z',
@@ -260,7 +260,7 @@ describe('Build', function() {
                   'version': '1.0.1'
               }
           ],
-          'runtime': 'node10',
+          'runtime': 'node14',
           'status': 'building',
           'date_created': '2018-11-10T20:00:00Z',
           'date_updated': '2018-11-10T20:00:00Z',