+++ spec/integration/rest/verify/v2/service.spec.js
@@ -78,6 +78,7 @@ describe('Service', function() {
               'code_length': 3,
               'skew': 2
           },
+          'default_template_sid': 'HJaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
           'date_created': '2015-07-30T20:00:00Z',
           'date_updated': '2015-07-30T20:00:00Z',
           'url': 'https://verify.twilio.com/v2/Services/VAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
@@ -151,6 +152,7 @@ describe('Service', function() {
               'code_length': null,
               'skew': null
           },
+          'default_template_sid': 'HJaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
           'date_created': '2015-07-30T20:00:00Z',
           'date_updated': '2015-07-30T20:00:00Z',
           'url': 'https://verify.twilio.com/v2/Services/VAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
@@ -249,6 +251,7 @@ describe('Service', function() {
                       'code_length': null,
                       'skew': null
                   },
+                  'default_template_sid': 'HJaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'date_created': '2015-07-30T20:00:00Z',
                   'date_updated': '2015-07-30T20:00:00Z',
                   'url': 'https://verify.twilio.com/v2/Services/VAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
@@ -305,6 +308,7 @@ describe('Service', function() {
                       'code_length': null,
                       'skew': null
                   },
+                  'default_template_sid': 'HJaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'date_created': '2015-07-30T20:00:00Z',
                   'date_updated': '2015-07-30T20:00:00Z',
                   'url': 'https://verify.twilio.com/v2/Services/VAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
@@ -366,6 +370,7 @@ describe('Service', function() {
                       'code_length': null,
                       'skew': null
                   },
+                  'default_template_sid': 'HJaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'date_created': '2015-07-30T20:00:00Z',
                   'date_updated': '2015-07-30T20:00:00Z',
                   'url': 'https://verify.twilio.com/v2/Services/VAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
@@ -442,6 +447,7 @@ describe('Service', function() {
                       'code_length': null,
                       'skew': null
                   },
+                  'default_template_sid': 'HJaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'date_created': '2015-07-30T20:00:00Z',
                   'date_updated': '2015-07-30T20:00:00Z',
                   'url': 'https://verify.twilio.com/v2/Services/VAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
@@ -516,6 +522,7 @@ describe('Service', function() {
               'code_length': 3,
               'skew': 2
           },
+          'default_template_sid': 'HJaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
           'date_created': '2015-07-30T20:00:00Z',
           'date_updated': '2015-07-30T20:00:00Z',
           'url': 'https://verify.twilio.com/v2/Services/VAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',